Do not call methods unnecessarily
When using an array like this, you are wasting processing time to build menus that will never be used.  Add to this, if the menu option does any sort of processing, you will do that processing, even if the option wasn't selected.

This is a stop gap way of handling this.  It's not very SOLID, but it's a start.